Preparation for Assignment 03
In order to continue from the previous assignment, I will start this exercise
from the teacher's version of the CanRack class. It will make things easier
in the long run. I also had to get rid of my Bin class and pass my variables
to the he Can class, this being another requirement for this exercise.

For now, it should be easier to represent cans in the rack using integers
instead of objects. This will let me add control methods using enumerate
variables.