Done with Exercise 2.2, working on 2.3
I have put in practice the idea of building from simpler code. The more
code the more possibility of error?

The new CanRack class requires a Enum Implementation and the new
PurchasePrice class requires a Decimal implementation.

The programmer is required to support both newer and legacy code,
Enumerations have made debugging easier by limiting the number of
inputs.

Now, I'm at a point from Exercise 2.3 where taking a decimal input to
transform it to Enum is not working as intended. There is a logic error
not giving any output.