Working on 2.6
It might be a better idea that the users of legacy string methods go
through a validation process. If the strings are valid then they will
be transformed into ENUM.

The way I had it before was ENUM is transformed to string and then send
to the string method.

The intention of this change is to make migration easier, meaning there
will be no effect to the ENUM methods in case I erase the string
ones.