Assignment 4 done
I was able to considerably simplify the code by making use of
enum methods such as .Parse .IsDefined .TryParse.

I also had the chance to experiment with out(write) and ref(read and
write) but didn't end up using it in my code, again in an effort to
simplify the flow.

My code supports new and legacy users of the classes. This is shown
in the main method while creating 'Coin', 'PurchasePrice' and 'CanRack'
objects using different constructors.