Done with 5.3
Getting ready for 5.4 Exception Handling.

In main, I had to create 1 coin for every denomination. These objects
are used to add coin denominations to the coinBox. If I create any
new temporary coins, then these have to reference back to the initial
coin objects, just so I can successfully use the Deposit or Withdraw
methods.

The machine is created with seed coins but it also keeps track of the coins
inserted so far on a separate list. This is just so people don't use my
machine to exchange their nickels for half-dollars.