Completed 5.4 Exceptions
It was good practice to rewrite some of my code using 'try' and
'catch' clauses, and in some cases this practice simplified my code.

Apart from exceptions, I created a 'Dictionary<Flavor,PurchasePrice>' and
this helped to avoid some repetition in my code. Regarding removing
an obj from a 'list<T>', I found out this is a hot topic, and the only way
to handle the task is by keeping track of the objects created on
separate list.