Branching out and Exercise 6 WPF
I started a new branch and included a few corrections to my
version of the Vending machine. The most important change is to
be able to remove an object from a list by matching a property
(instead of the object itself). This was possible by using 'LINQ' and
the type 'IEnumerable<T>'.

Apart from that, I integrated the console library developed
for the class. I also created the base GUI for the vending machine
but I still have to implement the event handlers.