Assignment 09 Completed
When creating a list of Snack objects, the parent's ToString method
will be called. In order to call the child's ToString() override, I
had to cast 'Snack s' to object (ex.(object(s).ToString()))

Somehow when downcasting to object and then calling ToString, it uses
the child's override instead of the parent's. Maybe because, this method
was initially overriden from the object class?

That would probably mean overriding another child method (apart from the
Object.ToString) will not produce the same results when casted to object